+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@codistan-isb/ranchers-api-plugin-orders",
   "description": "Orders plugin for the Reaction API",
-  "version": "7.3.9",
+  "version": "7.3.10",
   "main": "index.js",
   "type": "module",
   "engines": {
